requirements.txt
blinker==1.9.0
click==8.2.1
dependency-injector==4.48.1
Flask==3.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pytest==8.4.1
Werkzeug==3.0.1
zen-engine==0.49.1
